requirements.txt
Flask==2.3.3
flask-bcrypt==1.0.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
